ssb-ahoy work (tidying)
