+++ package.json
@@ -98,10 +98,10 @@
     "ssb-chess-mithril": "1.0.10",
     "ssb-config": "^3.2.3",
     "ssb-ebt": "^5.5.2",
-    "ssb-friends": "^3.1.12",
     "ssb-friend-pub": "^1.0.5",
+    "ssb-friends": "^3.1.12",
     "ssb-invite": "^2.0.3",
-    "ssb-legacy-conn": "1.0.23",
+    "ssb-legacy-conn": "^1.0.23",
     "ssb-meme": "^1.0.4",
     "ssb-mentions": "^0.5.0",
     "ssb-mutual": "^0.1.0",
@@ -109,7 +109,7 @@
     "ssb-private": "^0.2.3",
     "ssb-query": "^2.1.0",
     "ssb-ref": "^2.13.6",
-    "ssb-replicate": "^1.0.4",
+    "ssb-replicate": "^1.2.4",
     "ssb-search": "^1.1.2",
     "ssb-server": "^14.1.9",
     "ssb-sort": "^1.1.0",
